error: resolve error that load user when delete user
